requirements.txt
Django==4.2.14
gunicorn==22.0.0
psycopg2-binary==2.9.9
django-environ==0.11.0
whitenoise==6.6.0